+++ azure-mgmt-containerservice/azure/mgmt/containerservice/v2017_07_01/models/orchestrator_version_profile_py3.py
@@ -22,10 +22,11 @@ class OrchestratorVersionProfile(Model):
     :param orchestrator_version: Required. Orchestrator version (major, minor,
      patch).
     :type orchestrator_version: str
-    :param default: Required. Installed by default if version is not
-     specified.
+    :param default: Installed by default if version is not specified.
     :type default: bool
-    :param upgrades: Required. The list of available upgrade versions.
+    :param is_preview: Whether Kubernetes version is currently in preview.
+    :type is_preview: bool
+    :param upgrades: The list of available upgrade versions.
     :type upgrades:
      list[~azure.mgmt.containerservice.v2017_07_01.models.OrchestratorProfile]
     """
@@ -33,20 +34,20 @@ class OrchestratorVersionProfile(Model):
     _validation = {
         'orchestrator_type': {'required': True},
         'orchestrator_version': {'required': True},
-        'default': {'required': True},
-        'upgrades': {'required': True},
     }
 
     _attribute_map = {
         'orchestrator_type': {'key': 'orchestratorType', 'type': 'str'},
         'orchestrator_version': {'key': 'orchestratorVersion', 'type': 'str'},
         'default': {'key': 'default', 'type': 'bool'},
+        'is_preview': {'key': 'isPreview', 'type': 'bool'},
         'upgrades': {'key': 'upgrades', 'type': '[OrchestratorProfile]'},
     }
 
-    def __init__(self, *, orchestrator_type: str, orchestrator_version: str, default: bool, upgrades, **kwargs) -> None:
+    def __init__(self, *, orchestrator_type: str, orchestrator_version: str, default: bool=None, is_preview: bool=None, upgrades=None, **kwargs) -> None:
         super(OrchestratorVersionProfile, self).__init__(**kwargs)
         self.orchestrator_type = orchestrator_type
         self.orchestrator_version = orchestrator_version
         self.default = default
+        self.is_preview = is_preview
         self.upgrades = upgrades